# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LEVH API",
    "description": "Local-first memory layer for AI agents and humans",
    "version": "2.32.0"
  },
  "paths": {
...
    "/api/v1/memories/{memory_id}/review": {
      "post": {
        "summary": "Review Memory",
        "description": "Apply a spaced-repetition review decision to a memory.",
        "operationId": "review_memory_api_memories__memory_id__review_post_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewMemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}": {
      "get": {
        "summary": "Get Memory",
        "operationId": "get_memory_api_memories__memory_id__get_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Memory",
        "operationId": "update_memory_api_memories__memory_id__put_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Memory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Memory",
        "operationId": "delete_memory_api_memories__memory_id__delete_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/pin": {
      "patch": {
        "summary": "Pin Memory",
        "operationId": "pin_memory_api_memories__memory_id__pin_patch_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Memory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/reinforce": {
      "post": {
        "summary": "Reinforce Memory",
        "description": "Manually strengthen a memory \u2014 resets its decay clock and grows its\nstability, the same reinforcement that happens automatically on recall.",
        "operationId": "reinforce_memory_api_memories__memory_id__reinforce_post_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Memory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/feedback": {
      "post": {
        "summary": "Memory Feedback",
        "description": "Learn from recall outcomes: helpful=true reinforces the memory,\nhelpful=false weakens it so wrong/stale information fades out fast.",
        "operationId": "memory_feedback_api_memories__memory_id__feedback_post_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Memory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/redact": {
      "post": {
        "summary": "Redact Memory",
        "description": "Strip secrets from an already-stored memory in place, recorded\nauditably in its metadata's redaction_history.",
        "operationId": "redact_memory_api_memories__memory_id__redact_post_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedactMemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/purge": {
      "post": {
        "summary": "Purge Memory",
        "description": "Hard-delete a memory across every layer and verify nothing survives.\nPinned memories are purged too \u2014 this is a deliberate human action.",
        "operationId": "purge_memory_api_memories__memory_id__purge_post_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeMemoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/forgetting-curve": {
      "get": {
        "summary": "Get Forgetting Curve",
        "description": "Predicted retention curve for a memory \u2014 powers the 'memory strength'\nvisualization in the dashboard's detail drawer.",
        "operationId": "get_forgetting_curve_api_memories__memory_id__forgetting_curve_get_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgettingCurveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/trust": {
      "get": {
        "summary": "Get Memory Trust",
        "description": "Provenance/trust breakdown for a memory \u2014 explainable, deterministic,\nNOT truth, and independent of H-score recall ranking.",
        "operationId": "get_memory_trust_api_memories__memory_id__trust_get_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustBreakdownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/related": {
      "get": {
        "summary": "Related Memories",
        "description": "Memories most similar to this one \u2014 the 'related memories' graph edge,\ncomputed live from embeddings. Powers 'see also' in the detail drawer.",
        "operationId": "related_memories_api_memories__memory_id__related_get_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelatedMemoriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memories/{memory_id}/score-breakdown": {
      "get": {
        "summary": "Get Score Breakdown",
        "description": "Return H(x,\u03c8) score breakdown for a specific memory + query pair.\n\nIf query is empty, memory.content is used as the default query so the\nbreakdown reflects self-similarity (baseline score for the memory).\n\nDelegates to ``engine.score_breakdown`` \u2014 the single source of truth for\ncosine/decay/scoring math \u2014 and reshapes its result into the public API\nschema. Contract: 404 unknown memory, 400 memory without embedding or\nwith a mismatched embedding dimension.",
        "operationId": "get_score_breakdown_api_memories__memory_id__score_breakdown_get_v1",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreBreakdownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeedbackRequest": {
        "properties": {
          "helpful": {
            "type": "boolean",
            "title": "Helpful"
          }
        },
        "type": "object",
        "required": [
          "helpful"
        ],
        "title": "FeedbackRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "MemoryDeleteResponse"
      },
      "MemoryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "memory_type": {
            "$ref": "#/components/schemas/MemoryType",
            "default": "short_term"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "importance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Importance",
            "default": 0.5
          },
          "frequency": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Frequency",
            "default": 1
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": false
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "hscore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hscore"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "accessed_at": {
            "type": "string",
            "title": "Accessed At"
          },
          "decay_factor": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Decay Factor",
            "default": 1.0
          },
          "stability_hours": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Stability Hours",
            "description": "This memory's own half-life in hours \u2014 how long until it decays to 50% relevance since last access. Grows every time the memory is recalled or explicitly reinforced (spaced-repetition style), so frequently-used memories become durable while unused ones fade.",
            "default": 168.0
          },
          "recall_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Recall Count",
            "description": "Times this memory has been reinforced by recall.",
            "default": 0
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentOut"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MemoryOut",
        "description": "A memory as the HTTP surface returns it: embedding dropped (large and\nuseless to a client - similarity is computed server-side) and attachments\njoined where the route does that."
      },
...
      "PinRequest": {
        "properties": {
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": true
          }
        },
        "type": "object",
        "title": "PinRequest"
      },
...
      "PurgeMemoryResponse": {
        "properties": {
          "memory_id": {
            "type": "string",
            "title": "Memory Id"
          },
          "existed": {
            "type": "boolean",
            "title": "Existed"
          },
          "purged": {
            "type": "boolean",
            "title": "Purged"
          },
          "residue": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Residue"
          }
        },
        "type": "object",
        "required": [
          "memory_id",
          "existed",
          "purged",
          "residue"
        ],
        "title": "PurgeMemoryResponse",
        "description": "``POST /api/memories/{id}/purge`` - the purge postcondition audit."
      },
...
      "RedactMemoryResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "redacted": {
            "type": "boolean",
            "title": "Redacted"
          },
          "secrets": {
            "items": {},
            "type": "array",
            "title": "Secrets"
          },
          "memory_id": {
            "type": "string",
            "title": "Memory Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "redacted",
          "memory_id"
        ],
        "title": "RedactMemoryResponse",
        "description": "``POST /api/memories/{id}/redact``."
      },
...
      "ReviewMemoryResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "memory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Id"
          },
          "review_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Count"
          },
          "review_due_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Due At"
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ReviewMemoryResponse",
        "description": "The result of a spaced-repetition review decision."
      },
...
      "ReviewRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "snooze_days": {
            "type": "integer",
            "title": "Snooze Days",
            "default": 7
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ReviewRequest"
      },
...
      "UpdateRequest": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "importance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned"
          }
        },
        "type": "object",
        "title": "UpdateRequest"
      },
...
    }
  }
}